add new commands and subcommands

+++ cmd/cli/stop.go
@@ -15,30 +15,31 @@ func initStopCommand() *cobra.Command {
 		Use:   "stop",
 		Short: "Stop kntrl daemon",
 		Run: func(cmd *cobra.Command, args []string) {
-			data, err := os.ReadFile(pidfile)
-			if err != nil {
-				qwe(127, err, "failed to read pidfile")
-			}
+			if daemon {
+				data, err := os.ReadFile(pidfile)
+				if err != nil {
+					qwe(127, err, "failed to read pidfile")
+				}
 
-			pid, err := strconv.Atoi(string(data))
-			if err != nil {
-				qwe(127, err, "failed to convert pid")
-			}
+				pid, err := strconv.Atoi(string(data))
+				if err != nil {
+					qwe(127, err, "failed to convert pid")
+				}
 
-			process, err := os.FindProcess(pid)
-			if err != nil {
-				qwe(127, err, "failed to find process id -- is kntrl running?")
-			}
+				process, err := os.FindProcess(pid)
+				if err != nil {
+					qwe(127, err, "failed to find process id -- is kntrl running?")
+				}
 
-			if err := process.Kill(); err != nil {
-				qwe(127, err, "failed to kill the process id")
-			}
+				if err := process.Kill(); err != nil {
+					qwe(127, err, "failed to kill the process id")
+				}
 
-			if err := os.Remove(pidfile); err != nil {
-				qwe(127, err, "failed to remove pidfile")
+				if err := os.Remove(pidfile); err != nil {
+					qwe(127, err, "failed to remove pidfile")
+				}
+				fmt.Printf("Process id [%s] stopped", string(data))
 			}
-
-			fmt.Printf("Process id [%s] stopped", string(data))
 			reporter.LoadAndPrint()
 		},
 	}
